Set up new endpoint for discovery (#136)
Since the /api/discovery_info endpoint has been deprecated, we lost a
way to retrieve a unique identifier from the remote instance. This
change introduces /api/remote_homeassistant/discovery containing the
information we are missing. The downside is that the component now must
be installed on the remote instance as well. The upside however is that
we now can add any new endpoint we need for additional features in the
future, which is good.

Fixes #133